google_function/pyproject.toml
[tool.poetry]
name = "news-feed"
version = "0.1.0"
description = ""
authors = ["Thomas Cadet <thomas.cadet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jupyter = "^1.0.0"
feedparser = "^6.0.11"
pandas = "^2.2.2"
google-cloud-storage = "^2.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
